requirement.txt
git+https://github.com/UKPLab/sentence-transformers.git
beir
gdown
torch
pyprojroot
transformers-interpret
accelerate